+++ MaestroSimulation/Plugins/ObjectiveSystem/Source/ObjectiveSystem/Public/ObjectiveWorldSubsystem.h
@@ -5,7 +5,7 @@
 #include "ObjectiveWorldSubsystem.generated.h"
 
 // The signature for the broadcast
-DECLARE_DYNAMIC_MULTICAST_DELEGATE_TwoParams(FOnObjectiveEvent, FName, EventTag, int32, Value);
+DECLARE_DYNAMIC_MULTICAST_DELEGATE_TwoParams(FOnObjectiveEvent, FName, EventTag, float, Value);
 
 /**
  * A global router for objective events.
@@ -22,5 +22,5 @@ public:
 
 	// Called by any actor involved in the quest system (e.g. bristle blaster, drone)
 	UFUNCTION(BlueprintCallable, Category = "Objective System", meta=(WorldContext="WorldContextObject"))
-	void BroadcastObjectiveEvent(FName Tag, int32 Value = 1);
+	void BroadcastObjectiveEvent(FName Tag, float Value = 1);
 };
